Move the Guest class from base to steps.provision
Here it makes more sense as it is related to provision classes.
