+++ tmt/__init__.py
@@ -1,5 +1,6 @@
 """ Test Management Tool """
 
-from tmt.base import Tree, Test, Plan, Story, Run, Guest, Result
+from tmt.base import Tree, Test, Plan, Story, Run, Result
+from tmt.steps.provision import Guest
 
 __all__ = ['Tree', 'Test', 'Plan', 'Story', 'Run', 'Guest', 'Result']
