+++ tmt/base.py
@@ -7,8 +7,6 @@ import fmf
 import yaml
 import click
 import pprint
-import random
-import string
 import subprocess
 
 import tmt.steps
@@ -908,197 +906,6 @@ class Run(tmt.utils.Common):
 
         # Give the final summary, remove workdir, handle exit codes
         self.finish()
-
-
-class Guest(tmt.utils.Common):
-    """
-    Guest environment provisioned for test execution
-
-    The following keys are expected in the 'data' dictionary::
-
-        guest ...... hostname or ip address
-        user ....... user name to log in
-        key ........ private key
-        password ... password
-
-    These are by default imported into instance attributes.
-    """
-
-    # Supported keys (used for import/export to/from attributes)
-    _keys = ['guest', 'user', 'key', 'password']
-
-    def __init__(self, data, name=None, parent=None):
-        """ Initialize guest data """
-        super().__init__(parent, name)
-        self.load(data)
-
-    def _random_name(self):
-        """ Generate a random name """
-        return ''.join(random.choices(string.ascii_letters, k=16))
-
-    def _ssh_guest(self):
-        """ Return user@guest """
-        return f'{self.user}@{self.guest}'
-
-    def _ssh_options(self, join=False):
-        """ Return common ssh options (list or joined) """
-        options = [
-            '-oStrictHostKeyChecking=no',
-            '-oUserKnownHostsFile=/dev/null',
-            ]
-        if self.key:
-            options.extend(['-i', self.key])
-        return ' '.join(options) if join else options
-
-    def _ssh_command(self, join=False):
-        """ Prepare an ssh command line for execution (list or joined) """
-        command = ['sshpass', f'-p{self.password}'] if self.password else []
-        command += ['ssh'] + self._ssh_options()
-        return ' '.join(command) if join else command
-
-    def load(self, data):
-        """ Load guest data for easy access """
-        for key in self._keys:
-            setattr(self, key, data.get(key))
-
-    def save(self):
-        """ Save guest data for future wake up """
-        data = dict()
-        for key in self._keys:
-            value = getattr(self, key)
-            if value is not None:
-                data[key] = value
-        return data
-
-    def wake(self):
-        """ Wake up the guest """
-        self.debug(f"Doing nothing to wake up guest '{self.guest}'.")
-
-    def start(self):
-        """ Start the guest """
-        self.debug(f"Doing nothing to start guest '{self.guest}'.")
-
-    def details(self):
-        """ Show guest details such as distro and kernel """
-        # Skip distro & kernel check in dry mode
-        if self.opt('dry'):
-            return
-        # Distro
-        try:
-            distro = self.execute(
-                'cat /etc/redhat-release', dry=True)[0].strip()
-        except tmt.utils.RunError:
-            try:
-                distro = self.execute('cat /etc/lsb-release')[0].strip()
-                distro = re.search('DESCRIPTION="(.*)"', distro).group(1)
-            except (tmt.utils.RunError, AttributeError):
-                distro = None
-        if distro:
-            self.info('distro', distro, 'green')
-        # Kernel
-        kernel = self.execute('uname -r', dry=True)[0].strip()
-        self.verbose('kernel', kernel, 'green')
-
-    def _ansible_verbosity(self):
-        """ Prepare verbose level based on the --debug option count """
-        if self.opt('debug') < 3:
-            return ''
-        else:
-            return ' -' + (self.opt('debug') - 2) * 'v'
-
-    def _ansible_summary(self, output):
-        """ Check the output for ansible result summary numbers """
-        if not output:
-            return
-        keys = 'ok changed unreachable failed skipped rescued ignored'.split()
-        for key in keys:
-            matched = re.search(rf'^.*\s:\s.*{key}=(\d+).*$', output, re.M)
-            if matched and int(matched.group(1)) > 0:
-                tasks = fmf.utils.listed(matched.group(1), 'task')
-                self.verbose(key, tasks, 'green')
-
-    def _ansible_playbook_path(self, playbook):
-        """ Prepare full ansible playbook path """
-        # Playbook paths should be relative to the metadata tree root
-        self.debug(f"Applying playbook '{playbook}' on guest '{self.guest}'.")
-        playbook = os.path.join(self.parent.plan.run.tree.root, playbook)
-        self.debug(f"Playbook full path: '{playbook}'", level=2)
-        return playbook
-
-    def _export_environment(self, execute_environment=None):
-        """ Prepare shell export of environment variables """
-        # Prepare environment variables so they can be correctly passed
-        # to ssh's shell. Create a copy to prevent modifying source.
-        environment = dict()
-        environment.update(execute_environment or dict())
-        # Plan environment and variables provided on the command line
-        # override environment provided to execute().
-        environment.update(self.parent.plan.environment)
-        # Prepend with export and run as a separate command.
-        if not environment:
-            return ''
-        return 'export {}; '.format(
-            ' '.join(tmt.utils.shell_variables(environment)))
-
-    def ansible(self, playbook):
-        """ Prepare guest using ansible playbook """
-        playbook = self._ansible_playbook_path(playbook)
-        stdout, stderr = self.run(
-            f'stty cols {tmt.utils.OUTPUT_WIDTH}; ansible-playbook '
-            f'--ssh-common-args="{self._ssh_options(join=True)}" '
-            f'-e ansible_python_interpreter=auto'
-            f'{self._ansible_verbosity()} -i {self._ssh_guest()}, {playbook}')
-        self._ansible_summary(stdout)
-
-    def execute(self, command, **kwargs):
-        """
-        Execute command on the guest
-
-        command ....... string or list of command arguments
-        environment ... dictionary with environment variables
-        """
-
-        # Prepare the export of environment variables
-        environment = self._export_environment(kwargs.get('env', dict()))
-
-        # Change to given directory on guest if cwd provided
-        directory = kwargs.get('cwd', '')
-        if directory:
-            directory = f"cd '{directory}'; "
-
-        # Run in interactive mode if requested
-        interactive = ['-t'] if kwargs.get('interactive') else []
-
-        # Prepare command and run it
-        if isinstance(command, (list, tuple)):
-            command = ' '.join(command)
-        self.debug(f"Execute command '{command}' on guest '{self.guest}'.")
-        command = (
-            self._ssh_command() + interactive + [self._ssh_guest()] +
-            [f'{environment}{directory}{command}'])
-        return self.run(command, shell=False, **kwargs)
-
-    def push(self):
-        """ Push workdir to guest """
-        self.debug(f"Push workdir to guest '{self.guest}'.")
-        self.run(
-            f'rsync -Rrze "{self._ssh_command(join=True)}" --delete '
-            f'{self.parent.plan.workdir} {self._ssh_guest()}:/')
-
-    def pull(self):
-        """ Pull workdir from guest """
-        self.debug(f"Pull workdir from guest '{self.guest}'.")
-        self.run(
-            f'rsync -Rrze "{self._ssh_command(join=True)}" '
-            f'{self._ssh_guest()}:{self.parent.plan.workdir} /')
-
-    def stop(self):
-        """ Stop the guest """
-        self.debug(f"Doing nothing to stop guest '{self.guest}'.")
-
-    def remove(self):
-        """ Remove the guest (disk cleanup) """
-        self.debug(f"Doing nothing to remove guest '{self.guest}'.")
 
 
 class Result(object):
